google: Set auto_refresh_url only for offline
Only when we request an offline token do we actually get a refresh token
that can be used use the `auto_refresh_url` endpoint to refresh the
token. As such, when we get an online token instead (the default) we
shouldn't set `auto_refresh_url`.

This is the first part of the changes discussed in #144 and just ensures
that requests-oauthlib no longer tries to refresh an online token but
immediately raises a `TokenExpiredError`. This does not constitute an
API change since the resulting error is the same, but would at least
result in a more clear traceback.