test: Fix codecov reporting (#335)
By default, `coverage run` does not generate a `coverage.xml`, you have to call `coverage xml` for that. Since we're not doing that, `coverage.xml` doesn't exist so we don't end up upload code coverage.

It turns out however that doing this is unnecessary. The codecov action will do this by itself if it has to, as long as it finds a `.coverage` which coverage.py will write to by default.