Send specific kwargs to Oauth2CustomerBlueprint
Due to the additional setup in this function, it is preferable to use `make_gitlab_blueprint` instead of creating a `Oauth2CustomerBlueprint` directly. However, in the event of a self-hosted gitlab instance using self-signed certificates, these certificates cannot be verified and the redirection always results in an SSL error. Thus, one is unable to use `make_gitlab_blueprint` and must instead use the "raw" `Oauth2CustomerBlueprint` in order to pass `token_url_params=dict(verify=False)`.

This change is to add the ability to forward kwargs from `make_gitlab_blueprint` into `Oauth2CustomerBlueprint` for more customisationof functionality.